test(grpc): prune routine conversion tests; share fake_rpc_error fixture
Addresses PR #839 review feedback on test scope:
- Remove routine `_from_proto`/conversion tests (test_comment.py, test_unit.py,
  test_workspace.py, test_client_workspaces.py, and the get_unit happy-path) —
  they tested 1:1 conversion/wiring, not behavior.
- Keep only existing-test fixes and new business-logic/behavior tests
  (units flattening + validation, get_unit unknown-symbol→None contract +
  error propagation, error translation, resolve-path caching), each with a
  single-line behavior docstring.
- Consolidate the three duplicated `_FakeRpcError` classes into a single
  `fake_rpc_error` factory fixture in tests/conftest.py.
