+++ tests/core/test_grpc.py
@@ -209,18 +209,6 @@ def test_create_grpc_stub_factory_binds_stubs_to_one_shared_channel(monkeypatch)
     assert first_channel is stub_class.call_args_list[1].args[0]
 
 
-class _FakeRpcError(grpc.RpcError):
-    def __init__(self, code: grpc.StatusCode, details: str = "boom") -> None:
-        self._code = code
-        self._details = details
-
-    def code(self) -> grpc.StatusCode:
-        return self._code
-
-    def details(self) -> str:
-        return self._details
-
-
 @pytest.mark.parametrize(
     "code, expected",
     [
@@ -230,20 +218,20 @@ class _FakeRpcError(grpc.RpcError):
         (grpc.StatusCode.INVALID_ARGUMENT, NominalInvalidArgumentError),
     ],
 )
-def test_translate_grpc_errors_maps_known_status_codes(code, expected) -> None:
+def test_translate_grpc_errors_maps_known_status_codes(code, expected, fake_rpc_error) -> None:
     """Each mapped status code raises its dedicated NominalError subclass, chained to the RpcError."""
     with pytest.raises(expected) as exc_info:
         with translate_grpc_errors():
-            raise _FakeRpcError(code)
+            raise fake_rpc_error(code)
     assert isinstance(exc_info.value.__cause__, grpc.RpcError)
-    assert "boom" in str(exc_info.value)
+    assert "fake rpc error" in str(exc_info.value)
 
 
-def test_translate_grpc_errors_falls_back_to_base_for_unmapped_code() -> None:
+def test_translate_grpc_errors_falls_back_to_base_for_unmapped_code(fake_rpc_error) -> None:
     """An unmapped status code raises the base NominalError, not a subclass."""
     with pytest.raises(NominalError) as exc_info:
         with translate_grpc_errors():
-            raise _FakeRpcError(grpc.StatusCode.INTERNAL)
+            raise fake_rpc_error(grpc.StatusCode.INTERNAL)
     assert type(exc_info.value) is NominalError
